requirements.txt
cachetools
certifi
chardet
charset-normalizer
google-api-core
google-auth
google-cloud-core
google-cloud-translate
googleapis-common-protos
googletrans
grpcio
grpcio-status
h11
h2
hpack
hstspreload
httpcore
httpx
hyperframe
idna
protobuf
pyasn1
pyasn1-modules
requests
rfc3986
rsa
six
sniffio
urllib3
